move `BitPackingStrategy` to a more sensible place
